refactor.py -> refactor.pyi
